Implement new zkDbProcessor to mina-service
