--- packages/mina-service/src/circuit/index.ts
@@ -1,2 +1,0 @@
-export * from './circuit-factory';
-export * from './rollup-program';
