+++ packages/mina-service/src/domain/index.ts
@@ -1,2 +1,2 @@
-export * from './proof';
 export * from './zk-compile';
+export * from './rollup';
