+++ packages/mina-service/src/service/index.ts
@@ -1,4 +1,3 @@
 export * from './compile-service';
 export * from './offchain-rollup-service';
-export * from './task-service';
 export * from './transaction-service';
